Fetches linguist language list from GitHub
This pulls down a .yml file containing a list of languages known to
GitHub and parses it into our `Language` type. This required a bit
of preparation for the document first, as each section used the
language's name as the key, which we weren't able to access with
our YAML parser. So, on any line where we detect the language name
(as a key), we start a new document and move the name down to the
next line as the value in a key `name`, which frees it up to our
parser.