Collects commit data per language and writes it to file
We are now splitting apart the commits file into individual commits,
filtering out the commits by anyone but the desired authors based
on the authors and emails defined in the config, and summing up the
insertions and deletions for a given language per file per commit.

This data is being saved to a YAML file, but in the future it will
instead be used to create a data card image like what is currently
being used in this repo's `README`.