Allows users to list repos, including wildcards
Users can now provide an array of strings representing absolute
paths to the repos they want to include. These include the option
of adding a wildcard `*` at the end to include all directories
with the stated directory. Similarly, a config field to list the
same but for repos that should be excluded from consideration was
also added.